refactor: 도메인형 구조로 수정 - domain (#15)
