+++ src/main/java/com/smunity/api/domain/account/entity/User.java
@@ -1,4 +1,4 @@
-package com.smunity.api.data.entity;
+package com.smunity.api.domain.account.entity;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonProperty.Access;
