+++ src/main/java/com/smunity/api/domain/account/service/AccountService.java
@@ -1,7 +1,7 @@
-package com.smunity.api.service;
+package com.smunity.api.domain.account.service;
 
-import com.smunity.api.data.dto.SignInResultDto;
-import com.smunity.api.data.dto.SignUpResultDto;
+import com.smunity.api.domain.account.dto.SignInResultDto;
+import com.smunity.api.domain.account.dto.SignUpResultDto;
 
 
 public interface AccountService {
