+++ src/test/java/com/smunity/api/data/repository/UserRepositoryTest.java
@@ -1,6 +1,7 @@
 package com.smunity.api.data.repository;
 
-import com.smunity.api.data.entity.User;
+import com.smunity.api.domain.account.entity.User;
+import com.smunity.api.domain.account.repository.UserRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
